[bug-fix] * trainer.steps key is required and readded.
[refactor]
* callbacks file is split into separate ones.
* minor stuff.
